Fix WER regression from event-loop stage; harden alignment numerics
- sentence_detector: revert the single-forward SaT rewrite back to
  sat.split() with predict_proba only as the soft/hard-limit fallback;
  the rewrite changed split behavior and was the main WER regression.
  Keep the module-level lock serializing the shared SaT model across
  session threads (genuine missing-lock bug).
- audio_processor: restore the original PCM chunk sizing (single chunk
  up to max_bytes_per_sec). VAD event samples clamp to chunk boundaries,
  so the 1s-slice drain changed silence boundaries and shifted WER.
  Keep only the Silero call moved off the event loop via to_thread.
- simul_whisper: accumulate alignment mean/std running sums in float64;
  the E[x^2]-mean^2 form cancels catastrophically in fp32 on the
  strongly attended frame columns that decide the alignment argmax.
- whisper/model: force the manual fp32-softmax path for cross attention
  (allow_sdpa=False); SDPA's fp16 softmax shifts logits enough to flip
  borderline tokens.

WER after this batch: max-speed 0.0608/0.0592/0.0621 (baseline
0.0608-0.0612), realtime 0.0604/0.0737 (baseline 0.0625 n=1).

+++ whisperlivekit/simul_whisper/simul_whisper.py
@@ -573,8 +573,12 @@ class AlignAtt:
                     dim=1,
                 ).float()                              # (beam, n_align, rows, audio_len)
                 a_new = F.softmax(a_new, dim=-1)
-                row_sum = a_new.sum(dim=2)
-                row_sqsum = (a_new * a_new).sum(dim=2)
+                # Accumulate in float64: the E[x^2]-mean^2 variance form
+                # cancels catastrophically in fp32 exactly on the strongly
+                # attended frame columns (large mean, tiny variance) that
+                # decide the alignment argmax.
+                row_sum = a_new.sum(dim=2, dtype=torch.float64)
+                row_sqsum = (a_new * a_new).sum(dim=2, dtype=torch.float64)
                 if align_sum is None:
                     align_sum, align_sqsum = row_sum, row_sqsum
                 else:
@@ -583,8 +587,8 @@ class AlignAtt:
                 align_rows += a_new.shape[2]
                 mean = align_sum / align_rows
                 std = (align_sqsum / align_rows - mean * mean).clamp_min_(0).sqrt_()
-                last = (a_new[:, :, -1, :] - mean) / (std + 1e-8)
-                last = median_filter(last.unsqueeze(2), 7).squeeze(2)
+                last = (a_new[:, :, -1, :].double() - mean) / (std + 1e-8)
+                last = median_filter(last.float().unsqueeze(2), 7).squeeze(2)
                 attn_last_row = last.mean(dim=1)[:, :content_mel_len]
             else:
                 attn_last_row = torch.zeros(
